+++ test/integration/03-integration-peers.js
@@ -43,7 +43,7 @@ describe('03 - integration - peers', function () {
   }
 
   beforeEach('start cluster', function (done) {
-    this.timeout(4000);
+    this.timeout(20000);
 
     Promise.all([
       HappnerCluster.create(localInstanceConfig(1)),