+++ RELEASES.md
@@ -39,3 +39,8 @@
 4.1.0 2018-01-09
 ----------------
   - updated to happn-cluster 4.2.0 and happner-client 3.0.1
+
+5.0.0 2018-04-13
+----------------
+  - updated happn-cluster and happner-client and happner-2
+  - added tests for security update replication into cluster